TitanQuest2Fix: Added new FOV feature.
Added option to yaml to disable or enable
pillarbox fix.

Added new FOV feature. This allows you to set a
new FOV that the game will use instead. This
feature can be enabled or disabled. If enabled the
FOV provided in value shall be used by the game
or if disabled then the game will use its own
default FOV.

	modified:   TitanQuest2Fix.yml
	modified:   src/dllmain.cpp