TitanQuestFix: Fixed HUD feature
Previously for the HUD feature I hardcoded 1.125f
into the xmm0 register. I have now learned from
user feedback that is was wrong as they were
experiencing the HUD being completely messed up.
Through debugging I learned that what is in the
xmm0 is actually a ratio between the selected game
resolution and the physical desktop resolution.
For example if the game was set to 1920x1080 and
the desktop resolution was 3840x2160, the xmm0
regiter would show 0.5f (1080/2160). Now to apply
that 12.5% increase we multiply it by whatever is
in the xmm0 register and add it back.

Bumped version to v1.2.1 to prep for next release.

	modified:   src/dllmain.cpp